+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -145,7 +145,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-24T10:23:38.562Z",
+    "2025-02-25T07:54:03.323Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -179,7 +179,7 @@ export const FileGetResponseBody$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-24T10:23:38.562Z"))
+  created: z.date().default(() => new Date("2025-02-25T07:54:03.323Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {