+++ packages/orq-rc/RELEASES.md
@@ -529,3 +529,13 @@ Based on:
 - [typescript v3.2.0-rc.11] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.11 - packages/orq-rc
+
+## 2025-02-25 07:55:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.502.0 (2.524.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.12] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.12 - packages/orq-rc